Refactor store naming by app-state-store
