+++ pruebas/01-reading-list/oscarprdev/src/core/state/application/state.usecase.ts
@@ -1,4 +1,4 @@
-import { appState } from '../../store/store'
+import { appState } from '../../store/app-state-store'
 import type { GlobalState } from '../../types'
 import type { StatePorts } from './state.ports'
 import type { DefaultBooks } from './state.types'
